Load location from metadata
