+++ Brand/Database.swift
@@ -26,4 +26,4 @@ import Foundation
 // Database Realm
 //
 let databaseName                    = "nextcloud.realm"
-let databaseSchemaVersion: UInt64   = 302
+let databaseSchemaVersion: UInt64   = 303
