+++ iOSClient/Utility/NCUtility.swift
@@ -797,135 +797,21 @@ class NCUtility: NSObject {
         
         return titleView
     }
-    
-    public struct ExifData {
-        var colorModel: String?
-        var pixelWidth: Double?
-        var pixelHeight: Double?
-        var dpiWidth: Int?
-        var dpiHeight: Int?
-        var depth: Int?
-        var orientation: Int?
-        var apertureValue: Double?
-        var exposureValue: Int?
-        var shutterSpeedApex: Double?
-        var iso: Int?
-        var lensLength: Int?
-        var brightnessValue: String?
-        var dateTimeDigitized: String?
-        var dateTimeOriginal: String?
-        var offsetTime: String?
-        var offsetTimeDigitized: String?
-        var offsetTimeOriginal: String?
-        var make: String?
-        var model: String?
-        var software: String?
-        var tileLength: Double?
-        var tileWidth: Double?
-        var xResolution: Double?
-        var yResolution: Double?
-        var altitude: String?
-        var destBearing: String?
-        var hPositioningError: String?
-        var imgDirection: String?
-        var latitude: Double?
-        var longitude: Double?
-        var speed: Double?
-        var location: String?
-        var lensModel: String?
-        var date: Date?
-    }
-    
-    func getExif(metadata: tableMetadata, completion: @escaping (ExifData?) -> Void) {
-        if metadata.classFile != "image" || !CCUtility.fileProviderStorageExists(metadata) {
-            print("Storage exists or file is not an image")
-            return
-        }
-        
-        let url = URL(fileURLWithPath: CCUtility.getDirectoryProviderStorageOcId(metadata.ocId, fileNameView: metadata.fileNameView))
-        
-        guard let originalSource = CGImageSourceCreateWithURL(url as CFURL, nil) else {
-            print("Storage exists or file is not an image")
-            return
-        }
-        
-        guard let fileProperties = CGImageSourceCopyProperties(originalSource, nil) as NSDictionary?,
-              let fileSizeNumber = fileProperties[kCGImagePropertyFileSize] as? NSNumber,
-              let imageProperties = CGImageSourceCopyPropertiesAtIndex(originalSource, 0, nil) as NSDictionary? else {
-            print("Could not get file properties")
-            return
-        }
-        
-        var data = ExifData()
-        
-        data.colorModel = imageProperties[kCGImagePropertyColorModel] as? String
-        data.pixelWidth = imageProperties[kCGImagePropertyPixelWidth] as? Double
-        data.pixelHeight = imageProperties[kCGImagePropertyPixelHeight] as? Double
-        data.dpiWidth = imageProperties[kCGImagePropertyDPIWidth] as? Int
-        data.dpiHeight = imageProperties[kCGImagePropertyDPIHeight] as? Int
-        data.depth = imageProperties[kCGImagePropertyDepth] as? Int
-        data.orientation = imageProperties[kCGImagePropertyOrientation] as? Int
-        
-        if let tiffData = imageProperties[kCGImagePropertyTIFFDictionary] as? NSDictionary {
-            data.make = tiffData[kCGImagePropertyTIFFMake] as? String
-            data.model = tiffData[kCGImagePropertyTIFFModel] as? String
-            data.software = tiffData[kCGImagePropertyTIFFSoftware] as? String
-            data.tileLength = tiffData[kCGImagePropertyTIFFTileLength] as? Double
-            data.tileWidth = tiffData[kCGImagePropertyTIFFTileWidth] as? Double
-            data.xResolution = tiffData[kCGImagePropertyTIFFXResolution] as? Double
-            data.yResolution = tiffData[kCGImagePropertyTIFFYResolution] as? Double
-            
-            let dateTime = tiffData[kCGImagePropertyTIFFDateTime] as? String
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateFormat = "yyyy:MM:dd HH:mm:ss"
-            data.date = dateFormatter.date(from: dateTime ?? "")
-        }
-        
-        if let exifData = imageProperties[kCGImagePropertyExifDictionary] as? NSDictionary {
-            
-            data.apertureValue = exifData[kCGImagePropertyExifFNumber] as? Double
-            data.exposureValue = exifData[kCGImagePropertyExifExposureBiasValue] as? Int
-            data.shutterSpeedApex = exifData[kCGImagePropertyExifShutterSpeedValue] as? Double
-            data.iso = (exifData[kCGImagePropertyExifISOSpeedRatings] as? Array<Int>)?[0]
-            data.lensLength = exifData[kCGImagePropertyExifFocalLenIn35mmFilm] as? Int
-            data.brightnessValue = exifData[kCGImagePropertyExifBrightnessValue] as? String
-            data.dateTimeDigitized = exifData[kCGImagePropertyExifDateTimeDigitized] as? String
-            data.dateTimeOriginal = exifData[kCGImagePropertyExifDateTimeOriginal] as? String
-            data.offsetTime = exifData[kCGImagePropertyExifOffsetTime] as? String
-            data.offsetTimeDigitized = exifData[kCGImagePropertyExifOffsetTimeDigitized] as? String
-            data.offsetTimeOriginal = exifData[kCGImagePropertyExifOffsetTimeOriginal] as? String
-            data.lensModel = exifData[kCGImagePropertyExifLensModel] as? String
-        }
-        
-        if let gpsData = imageProperties[kCGImagePropertyGPSDictionary] as? NSDictionary {
-            data.altitude = gpsData[kCGImagePropertyGPSAltitude] as? String
-            data.destBearing = gpsData[kCGImagePropertyGPSDestBearing] as? String
-            data.hPositioningError = gpsData[kCGImagePropertyGPSHPositioningError] as? String
-            data.imgDirection = gpsData[kCGImagePropertyGPSImgDirection] as? String
-            data.latitude = gpsData[kCGImagePropertyGPSLatitude] as? Double
-            data.longitude = gpsData[kCGImagePropertyGPSLongitude] as? Double
-            data.speed = gpsData[kCGImagePropertyGPSSpeed] as? Double
-        }
-        
-        if let latitude = data.latitude, let longitude = data.longitude {
-            let geocoder = CLGeocoder()
-            let llocation = CLLocation(latitude: latitude, longitude: longitude)
-            
-            
-            geocoder.reverseGeocodeLocation(llocation) { placemarks, error in
-                if error == nil, let placemark = placemarks?.last {
-                    let locationComponents: [String] = [placemark.thoroughfare, placemark.postalCode, placemark.locality, placemark.administrativeArea, placemark.country]
-                        .compactMap{$0}
-                    
-                    data.location = locationComponents.joined(separator: ", ")
-                    completion(data)
-                }
-                
-                completion(data)
-            }
-        }
 
-        completion(data)
+    func getLocation(latitude: Double, longitude: Double, completion: @escaping (String?) -> Void) {
+        let geocoder = CLGeocoder()
+        let llocation = CLLocation(latitude: latitude, longitude: longitude)
+
+        geocoder.reverseGeocodeLocation(llocation) { placemarks, error in
+            if error == nil, let placemark = placemarks?.last {
+                let locationComponents: [String] = [placemark.thoroughfare, placemark.postalCode, placemark.locality, placemark.administrativeArea, placemark.country]
+                    .compactMap{$0}
+
+                completion(locationComponents.joined(separator: ", "))
+            }
+
+            completion(nil)
+        }
     }
 }
